Do not completely remove ~/.rustup/tmp and download
As discussed in #2430, it is better to keep the top-level directory.
It might be a mountpoint, or a symlink which the user wants to keep,
or have unusual permissions.

I wasn't able to find an implementation of "remove contents of this
directory but not the directory itself", so I have implemeted one.
Maybe this should be in some other crate somewhere but it's not that
huge so I just put it here.

This change changes the behaviour of `delete_dir_contents` to actually
just remove the contents, like the name suggests.  There are two call
sites: one to delete tmp and one to delete download.  This is correct
behaviour for both.

Closes #2430.
Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>